Extract the shared status report screen and sharing helpers
